@@ -1,10 +1,9 @@
 ﻿using NUnit.Framework;
 using Rebus.Tests.Contracts.Activation;
 
-namespace Rebus.CastleWindsor.Tests
+namespace Rebus.CastleWindsor.Tests;
+
+[TestFixture]
+public class CastleWindsorContainerTests : ContainerTests<CastleWindsorActivationContext>
 {
-    [TestFixture]
-    public class CastleWindsorContainerTests : ContainerTests<CastleWindsorActivationContext>
-    {
-    }
 }
