@@ -5,117 +5,117 @@ using System.Reflection;
 using Castle.MicroKernel.Registration;
 using Castle.MicroKernel.Registration.Lifestyle;
 using Castle.Windsor;
+using Rebus.Config;
 using Rebus.Handlers;
 // ReSharper disable UnusedMember.Global
 
-namespace Rebus.CastleWindsor
+namespace Rebus.CastleWindsor;
+
+/// <summary>
+/// Extension methods for making it easy to register Rebus handlers in your <see cref="WindsorContainer"/>
+/// </summary>
+public static class WindsorContainerExtensions
 {
     /// <summary>
-    /// Extension methods for making it easy to register Rebus handlers in your <see cref="WindsorContainer"/>
+    /// Uses an instance lifestyle where the instance is bound to (and thus will re-used across) the current Rebus transaction context
     /// </summary>
-    public static class WindsorContainerExtensions
+    public static ComponentRegistration<TService> PerRebusMessage<TService>(this LifestyleGroup<TService> lifestyleGroup) where TService : class
     {
-        /// <summary>
-        /// Uses an instance lifestyle where the instance is bound to (and thus will re-used across) the current Rebus transaction context
-        /// </summary>
-        public static ComponentRegistration<TService> PerRebusMessage<TService>(this LifestyleGroup<TService> lifestyleGroup) where TService : class
-        {
-            if (lifestyleGroup == null) throw new ArgumentNullException(nameof(lifestyleGroup));
-            return lifestyleGroup.Registration.LifestylePerRebusMessage();
-        }
+        if (lifestyleGroup == null) throw new ArgumentNullException(nameof(lifestyleGroup));
+        return lifestyleGroup.Registration.LifestylePerRebusMessage();
+    }
 
-        /// <summary>
-        /// Uses an instance lifestyle where the instance is bound to (and thus will re-used across) the current Rebus transaction context
-        /// </summary>
-        public static ComponentRegistration<TService> LifestylePerRebusMessage<TService>(this ComponentRegistration<TService> registration) where TService : class
-        {
-            if (registration == null) throw new ArgumentNullException(nameof(registration));
-            return registration.LifestyleScoped<RebusScopeAccessor>();
-        }
+    /// <summary>
+    /// Uses an instance lifestyle where the instance is bound to (and thus will re-used across) the current Rebus transaction context
+    /// </summary>
+    public static ComponentRegistration<TService> LifestylePerRebusMessage<TService>(this ComponentRegistration<TService> registration) where TService : class
+    {
+        if (registration == null) throw new ArgumentNullException(nameof(registration));
+        return registration.LifestyleScoped<RebusScopeAccessor>();
+    }
 
-        /// <summary>
-        /// Automatically picks up all handler types from the assembly containing <typeparamref name="THandler"/> and registers them in the container
-        /// </summary>
-        public static IWindsorContainer AutoRegisterHandlersFromAssemblyOf<THandler>(this IWindsorContainer container)
-        {
-            if (container == null) throw new ArgumentNullException(nameof(container));
+    /// <summary>
+    /// Automatically picks up all handler types from the assembly containing <typeparamref name="THandler"/> and registers them in the container
+    /// </summary>
+    public static IWindsorContainer AutoRegisterHandlersFromAssemblyOf<THandler>(this IWindsorContainer container)
+    {
+        if (container == null) throw new ArgumentNullException(nameof(container));
 
-            var assemblyToRegister = typeof(THandler).Assembly;
+        var assemblyToRegister = typeof(THandler).Assembly;
 
-            return RegisterAssembly(container, assemblyToRegister);
-        }
+        return RegisterAssembly(container, assemblyToRegister);
+    }
 
-        /// <summary>
-        /// Automatically picks up all handler types from the specified assembly and registers them in the container
-        /// </summary>
-        public static IWindsorContainer AutoRegisterHandlersFromAssembly(this IWindsorContainer container, Assembly assemblyToRegister)
-        {
-            if (container == null) throw new ArgumentNullException(nameof(container));
-            if (assemblyToRegister == null) throw new ArgumentNullException(nameof(assemblyToRegister));
+    /// <summary>
+    /// Automatically picks up all handler types from the specified assembly and registers them in the container
+    /// </summary>
+    public static IWindsorContainer AutoRegisterHandlersFromAssembly(this IWindsorContainer container, Assembly assemblyToRegister)
+    {
+        if (container == null) throw new ArgumentNullException(nameof(container));
+        if (assemblyToRegister == null) throw new ArgumentNullException(nameof(assemblyToRegister));
 
-            return RegisterAssembly(container, assemblyToRegister);
-        }
+        return RegisterAssembly(container, assemblyToRegister);
+    }
 
-        /// <summary>
-        /// Registers the given handler type under the implemented handler interfaces
-        /// </summary>
-        public static IWindsorContainer RegisterHandler<THandler>(this IWindsorContainer container) where THandler : IHandleMessages
-        {
-            if (container == null) throw new ArgumentNullException(nameof(container));
-            RegisterType(container, typeof(THandler), false);
-            return container;
-        }
+    /// <summary>
+    /// Registers the given handler type under the implemented handler interfaces
+    /// </summary>
+    public static IWindsorContainer RegisterHandler<THandler>(this IWindsorContainer container) where THandler : IHandleMessages
+    {
+        if (container == null) throw new ArgumentNullException(nameof(container));
+        RegisterType(container, typeof(THandler), false);
+        return container;
+    }
 
-        static IWindsorContainer RegisterAssembly(IWindsorContainer container, Assembly assemblyToRegister)
-        {
-            if (container == null) throw new ArgumentNullException(nameof(container));
-            if (assemblyToRegister == null) throw new ArgumentNullException(nameof(assemblyToRegister));
+    static IWindsorContainer RegisterAssembly(IWindsorContainer container, Assembly assemblyToRegister)
+    {
+        if (container == null) throw new ArgumentNullException(nameof(container));
+        if (assemblyToRegister == null) throw new ArgumentNullException(nameof(assemblyToRegister));
 
-            var typesToAutoRegister = GetConcreteTypes(assemblyToRegister)
-                .Select(type => new
-                {
-                    Type = type,
-                    ImplementedHandlerInterfaces = GetImplementedHandlerInterfaces(type).ToList()
-                })
-                .Where(a => a.ImplementedHandlerInterfaces.Any());
-
-            foreach (var type in typesToAutoRegister)
+        var typesToAutoRegister = GetConcreteTypes(assemblyToRegister)
+            .Select(type => new
             {
-                RegisterType(container, type.Type, true);
-            }
+                Type = type,
+                ImplementedHandlerInterfaces = GetImplementedHandlerInterfaces(type).ToList()
+            })
+            .Where(a => a.ImplementedHandlerInterfaces.Any());
 
-            return container;
-        }
-
-        static IEnumerable<Type> GetConcreteTypes(Assembly assemblyToRegister)
+        foreach (var type in typesToAutoRegister)
         {
-            return assemblyToRegister.GetTypes()
-                .Where(type => !type.IsInterface && !type.IsAbstract);
+            RegisterType(container, type.Type, true);
         }
 
-        static void RegisterType(IWindsorContainer container, Type typeToRegister, bool auto)
-        {
-            if (container == null) throw new ArgumentNullException(nameof(container));
-            if (typeToRegister == null) throw new ArgumentNullException(nameof(typeToRegister));
+        return container;
+    }
 
-            var implementedHandlerInterfaces = GetImplementedHandlerInterfaces(typeToRegister).ToArray();
+    static IEnumerable<Type> GetConcreteTypes(Assembly assemblyToRegister)
+    {
+        return assemblyToRegister.GetTypes()
+            .Where(type => !type.IsInterface && !type.IsAbstract);
+    }
 
-            if (!implementedHandlerInterfaces.Any()) return;
+    static void RegisterType(IWindsorContainer container, Type typeToRegister, bool auto)
+    {
+        if (container == null) throw new ArgumentNullException(nameof(container));
+        if (typeToRegister == null) throw new ArgumentNullException(nameof(typeToRegister));
 
-            container.Register(
-                Component.For(implementedHandlerInterfaces)
-                    .ImplementedBy(typeToRegister)
-                    .LifestyleTransient()
-                    .Named($"{typeToRegister.FullName} ({(auto ? "auto-registered" : "manually registered")})")
-                );
-        }
+        var implementedHandlerInterfaces = GetImplementedHandlerInterfaces(typeToRegister).ToArray();
 
-        static IEnumerable<Type> GetImplementedHandlerInterfaces(Type type)
-        {
-            if (type == null) throw new ArgumentNullException(nameof(type));
+        if (!implementedHandlerInterfaces.Any()) return;
 
-            return type.GetInterfaces()
-                .Where(i => i.IsGenericType && i.GetGenericTypeDefinition() == typeof (IHandleMessages<>));
-        }
+        container.Register(
+            Component.For(implementedHandlerInterfaces)
+                .ImplementedBy(typeToRegister)
+                .LifestyleTransient()
+                .Named($"{typeToRegister.FullName} ({(auto ? "auto-registered" : "manually registered")})")
+        );
+    }
+
+    static IEnumerable<Type> GetImplementedHandlerInterfaces(Type type)
+    {
+        if (type == null) throw new ArgumentNullException(nameof(type));
+
+        return type.GetInterfaces()
+            .Where(i => i.IsGenericType && i.GetGenericTypeDefinition() == typeof (IHandleMessages<>));
     }
 }
